api: Add ToolList creation endpoint
Add a new POST /api/lists/ route for creating new lists of tools.

The EditCommentFieldMixin previously found in
toolhub.apps.toolinfo.serializers is hoisted up to toolhub.serializers
so that it can be easily shared across sub-apps.

Basic permissions for add/change/delete/view operations on ToolList
models has been introduced in this patch. Currently the same
creator+admin permissions used for Tool models are applied to
change/delete operations for ToolList. We may revisit this in the future
if we decide to introduce community curated lists.

Bug: T271488
Change-Id: I9520c64b2fff01fd39e0d2df1c384be614ffd994