api: Add endpoints for list revisions
* GET /api/lists/{list_pk}/revisions/ - all revisions of a list
* GET /api/lists/{list_pk}/revisions/{id}/ - specific revision
* GET api/lists/{list_pk}/revisions/{id}/diff/{other_id}/ - diff
* PATCH /api/lists/{list_pk}/revisions/{id}/hide/ - hide a revision
* PATCH /api/lists/{list_pk}/revisions/{id}/patrol/ - mark patrolled
* PATCH /api/lists/{list_pk}/revisions/{id}/reveal/ - reveal hidden revision
* POST /api/lists/{list_pk}/revisions/{id}/revert/ - revert to revision
* POST /api/lists/{list_pk}/revisions/{id}/undo/{other_id}/ - undo

Hide/reveal actions require the same membership in the Oversighters or
Administrators user groups that is needed for the analogous toolinfo
endpoints. Patrol requires Patrollers or Administrators group
membership. Undo and revert require the calling user to be the creator of
the list or a member of the Administrators group.

During implementation it was discovered that the reversion library we
are using for taking snapshots of state for versions does not deal with
many-to-many relations really well. The workaround currently in use is
keeping a JSON blob in the list table which has the names of the list's
tools in parallel with the m2m models. This duplication is annoying, but
a relatively minor storage increase lets us better version things
without trying to fork or extend reversion (which looks to be a lot more
difficult than the workaround).

Bug: T271488
Change-Id: I7d892cdd9c793fb10a535872169091cd4b33eb01