backend: Add Content-Security-Policy header and report collector
Use the django-csp library to add a Content-Security-Policy header to
all responses with a report-uri pointing back to the backend server.

Our CSP policy is intended to be as closed as possible with the tooling
that we are using. For most directives, 'self' is the only allowed
origin. We are allowing 'unsafe-inline' for CSS due to the methods used
by the Vue framework to inject CSS into content. We are also allowing
data: uris for images.

When the backend is running in DEBUG=True mode we assume that we are in
a local testing environment and add additional origins for script-src,
font-src, and connect-src which allow communication with a Vue
development server running on localhost.

The /csp-report reporting endpoint attempts to filter out reports that
are likely to be unrelated to our application code or from old
user-agents with limited CSP reporting support by discarding the report
when:
* total size is >8KiB
* json payload is malformed
* "csp-report" key is missing
* "blocked-uri" or "source-file" begin with unexpected values
* "line-number" is present with a value <= 1

Reports that pass the filtering rules are logged at level "warning" to
the "csp-report" logging channel.

Bug: T278882
Change-Id: I85cc4d2de95d8556dff1deaca9b65ddaaa9be127