dev: Add FIREFOX_DEVTOOL_HACK setting
Add an environment variable and related Django configuration which can
be used to loosen the Content-Security-Policy emitted in a local
development environment so that the Vue devtools browser plugin will
load in Firefox.

This setting is off by default, but can be enabled by adding
`FIREFOX_DEVTOOL_HACK=true` to your local `.env` configuration file.

Bug: T287333
Change-Id: Ifd49ae10c152f6abd6e95fae5affad1edc91eef7