api: expand tests for toolinfo views
* Convert to use toolhub.tests.TestCase
* Add tests for edits by Oversighters
* Update ToolViewSet to support newly tested actions

The need for updates to ToolViewSet was discovered as a result of
writing the new tests. Previously Administrators and Oversighters were
blocked from update and delete of Tool instances by the `created_by`
filter applied by ToolViewSet.get_queryset(). Rather than doing a lot of
checking to see if the user should be able to see records created by
others and adjusting the filter() call we now rely on object permissions
to enforce access control.

Bug: T287286
Change-Id: Icea73be3900b97b866a9a90837dbda5df90b3691