api: expand test for lists views
* Convert to use toolhub.tests.TestCase
* Add tests for edits by Oversighters
* Update ToolListViewSet to support newly tested actions

The need for updates to ToolListViewSet was discovered as a result of
writing the new tests. Previously Administrators and Oversighters were
blocked from update and delete of Tool instances by the `created_by`
filter applied by ToolListViewSet.get_queryset(). Rather than doing a lot of
checking to see if the user should be able to see records created by
others and adjusting the filter() call we now rely on object permissions
to enforce access control.

Bug: T287286
Change-Id: Ic0a6f3deb03e0b06f407cfc11e95febd45a27d97