build: Tag CI built production images as "latest"
Add a secondary tag of "latest" to production images when pushing to the
repository. This tag is *not* intended for use in an actual production
environment, but will be useful in user acceptance testing environments.
Previously we envisioned a separate image for demos and UAT, but it
seems easy enough to control settings with environment variables and
entrypoint overrides in practice. (Also the config we wrote for that
image confused Blubber. See T287516 for more info.)

Bug: T278503
Change-Id: I197def61d4d292b133bec0c5d440c95d9734b4c3