docker-compose: Do not start prometheus by default
The docker-compose managed prometheus container is fun to play with, but
really not needed to develop Toolhub. This change introduces the use of
'profiles' (<https://docs.docker.com/compose/profiles/>) to mark both
the prometheus and oauth-client containers as separate from the default
application stack. This also allows us to simplify the Makefile slightly
by removing the DEFAULT_CONTAINERS setting which was previously used to
explicitly list the containers to start. The default behavior of
`docker-compose up` will now be used instead which is to start only the
containers with no profiles attached.

Change-Id: I4c3e29cd68a95bdac02f8e3352dd76dd5681de6b