search: rename toolinfo search index
Prefix the index name used on the Elasticsearch side with `toolhub_` so
that the index name is more descriptive for es admins.

NOTE: any local dev or demo deployments will need to rebuild the index
following application of this patch. The most direct way to do that in
a local dev environment would look something like:

    $ make web-shell
    $ poetry run python3 manage.py search_index --rebuild -f

Bug: T271376
Change-Id: I264eaf89c51e105d869d71d28600897a4d3c9e4d