ui: Upgrade nodejs libs
+ @casl/ability@5.4.0
+ @casl/vue@1.2.3
+ @intlify/eslint-plugin-vue-i18n@0.12.0
+ @vue/test-utils@1.2.2
+ @wikimedia/jsonschema-tools@0.10.4
+ @wikimedia/language-data@1.0.3
+ banana-i18n@2.2.0
+ core-js@3.16.1
+ eslint-plugin-vuetify@1.0.1
+ eslint@7.32.0
+ sass@1.37.5
+ sinon@11.1.2
+ sinon-chai@3.7.0
+ stylelint-config-wikimedia@0.11.1
+ swagger-client@3.13.7
+ vue-cli-plugin-vuetify@2.4.2
+ vue-i18n@8.25.0
+ vue-router@3.5.2
+ vue-template-compiler@2.6.14
+ vue@2.6.14
+ vuetify@2.5.8

swagger-client was found to not be fully respecting of semver, or at least
not considering dropping node v10 support a change requiring a major
version bump. `npm update` took it to 3.15.0, but testing found that
3.13.7 was the last version supported on node v10. As a result,
swagger-client has had it's version pin changed to "~3.13.7" to indicate
that only patch level changes should be considered safe for upgrade.

Bug: T288542
Change-Id: I1267541d577a2075ebae521ca94a0f59548f0d17