api: Add endpoints for managing owner-only authtokens
Enable Django REST Framework's TokenAuthentication mechanism and add
endpoints for managing a user's token.

* GET /api/user/authtoken/ - Get the user's authtoken
* POST /api/user/authtoken/ - Create the user's authtoken
* DELETE /api/user/authtoken/ - Delete the user's authtoken

An authtoken is an owner-only access mechanism for authenticating to the
Toolhub API. It is intended to be used by things interacting with the
API from a non-browser, non-interactive origin (like a bot). Each user
can have one valid authtoken at any given time. This token can be
deleted by the user at any time to invalidate it.

An authtoken is used by sending it as part of an "Authentication: Token
{token}" header when using the Toolhub API.

Bug: T288977
Change-Id: I288e785800468c6ef32c44872233d6793acfc384