crawler: catch ValidationError
Catch django.core.exceptions.ValidationError when upserting crawled
records into the database. This error state can occur when there is
a name collision between toolinfo records added via the API and records
gathered during a crawler run. The error is logged which will associate
it with the crawled URL and make it visible via the crawler run
reporting API endpoints.

Change-Id: I34179c03ed64934bdf5c56bbd39a6765f0b8aa4e