UI for editing / deleting a list
There are two buttons on the list info page on the top of the page, one
for editing and another for deleting. Clicking the Edit button invokes
the same form as used for list creation. While saving changes, the
comment dialog is invoked for describing the changes made. The list
is deleted when the Delete button is clicked. Notifications are
shown both when the list is edited and deleted.

Bug: T271491
Change-Id: I84240b276c6c0586eb489bb7a8a7f39dbad7247a