ui: Use named routes with :to= links
Use vue-router's named route syntax (<https://router.vuejs.org/api/#to>)
to describe links to other views rather than interpolated strings. Using
this interface should allow us to make changes to the router/index.js
routes without needing to find and fix all of the buttons and links
which reference a rendered route path. Updates will still be needed if
the parameters for the route change, but this is less likely that
deciding we want entity URLs to end in a `/` for example.

Change-Id: I296a7755bdfa455f6955bf11e0ad501aa5bc1a1c