ui[RegisterToolUrl]: Remove help sidebar
Remove the sidebar of help text from the "submit a toolinfo.json url"
tab. This was early content and not something that we have replicated on
any other screen. The message content itself was out of date ("In the
future, Toolhub will allow its users to submit their tools directly via
the interface.") as well.

Change-Id: I701aba2ad28c917614e0d3b1a2e6f04152ebadd8