oauth: Add SOCIAL_AUTH_PROXIES setting for production use
In a production deployment it may be necessary to use an HTTP proxy to
access meta.wikimedia.org for OAuth login. In theory setting an
HTTP_PROXY environment variable is all that is needed for this to work.
In practice however it is not clear if this is actually sufficient.
Social-core introduced a configuration setting in v3.4.0 that allows
explicit configuration of the proxies setting for the underlying
requests library.

Bug: T291447
Change-Id: I2bb0dbaca3f880eb2f99938a8553a44e8ce3e689