crawler: Add `--quiet` cli switch
Add a `--quiet` cli switch to our `manage.py crawl` command. When set
this will suppress printing of information about the crawler run when it
finishes. This is useful for things like running as an automated job
where stdout + stderr are collected as log events and not read by
humans.

Change-Id: Ieea0af3fa0d421da2b8b30c0943a42ae3704fc1b