backend: Add setting to toggle off urllib3 warnings
Add a URLLIB3_DISABLE_WARNINGS envvar setting that can be used to
disable urllib3 TLS warnings.

Urllib3 tries really hard to warn users about TLS problems. This is
a very nice feature of the library, but also a very spammy feature when
the developer is aware that things are not quite up to urllib3's
standards, but still operationally acceptable.

In the Wikimedia Foundation's production deployment outbound HTTP
requests are routed through some kind of proxy--either an envoy sidecar
or the 'url-downloader' proxy--and the initial hop connection to the
proxy itself is not TLS secured. This combination leads to
InsecureRequestWarning events from urllib3.

See also:
<https://urllib3.readthedocs.io/en/1.26.x/advanced-usage.html#ssl-warnings>

Bug: T292025
Change-Id: I31294566ee93c6191be932ca2dcf9e5a4bdd76e7