config: Do not force TLS for health checks
Add /healthz to the SECURE_REDIRECT_EXEMPT setting to exempt this
endpoint from 301 redirects generated by SecurityMiddleware for non-TLS
requests. The /healthz endpoint is used in Kubernetes deployments to
verify that the Django backend is running. These checks are done by the
kubelet from inside the Kubernetes cluster and do not traverse the CDN
edge which handles TLS termination.

Bug: T294072
Change-Id: Ic78ca40bfdbfd696402f43eb7e2e48d91b1c45ab