Feature: Auto-complete/lookahead search for adding tools
User Story:
As a Toolhub user,
I want to be able to find existing tools
So that I can add them to my list

Acceptance Criteria
* Given a tool is registered in toolhub, when I start typing the name
  of the tool, I expect to see a list of tools that match my search criteria
* Given a tool does not exist in toolhub, when I start typing the name
  of the tool, I expect to see a message stating no tool was found

Something worth mentioning is that I didn't make use of the already
available faceted search view.
Things like facets, multiple ordering and pagination are not neccessary
for this search implementation and given that this search view is
called several times before a user completes the input, it makes sense
to remove any bloat. What I did instead was to extract the basic
functionalities we need to get this autocomplte search working.

It is also worth mentioning that as a result of disabling pagination,
the search result is now limited to 10 items.

Bug: T294932
Change-Id: I16c5054711941f9ed25fe725140a45d142296efa