ui: add asterisk to signify field is required
Context:
Currently, for all forms, there is no way of knowing which fields are
required for submission until you interact with them. To improve this
experience for users, we should provide them context for which fields
they are required before filling out the form.

Acceptance Criteria:
Asterisk appears next to field labels that are required

Bug: T294941
Change-Id: Id04683c4f2ec56c83746793b3315b0f917ebae8c