ui: Expose GET /api/user/ promise via vuex
In this patch, the codebase was refactored so that the promise response
of GET /api/user/ from getUserInfo is stored as it is (in promise form),
and the user object from response.body also stored once the promise
have resolved.

Now every other component that depends on the state of user.js can choose
to wait for the promise to resolve through .then or not.

This patch was neccessary because we needed to avoid the race condition
between the initial call to GET /api/user/ endpoint made in App.vue's
created lifecycle hook and other attempts to use state gathered from the
endpoint in their own calls to other endpoints (see the linked bug below
on phabricator for more details).

Bug: T296263
Change-Id: I1e78418b3d875d0e1c159819d2efa3eb00dbca51