config: Add setting to disable Elasticsearch TLS cert verification
A new ES_VERIFY_CERTS env setting is added which can be used to disable
verification of TLS certificates when connecting to an Elasticsearch
cluster. This setting is not needed in production where we use an envoy
proxy to move TLS termination outside of the Django app. It is
potentially useful in a development setting or when (ab)using our
container as a maintenance environment connected to things via ssh
tunnels.

Bug: T290357
Bug: T276865
Change-Id: I655e488d2e396153ec36e5c6e5a91dad519dc75f