ui: Fix Locale Selection Bug in CrawlerHistory.vue
In the LineChart.vue component, the legend, titles, and dates shown on
the charts do not change when the locale is switched, unless the page
is reloaded or the user clicks on the pagination buttons.
this is because in crawlerHistory view, fillChartsData is responsible
for setting the data to be used by the chart
(including the string of the current locale) and this is called only
once, when crawlerHistory from the store changes.
This means that when the locale changes, fillChartsData is not called
so there is no way to update crawledUrlsChartData and
totalToolsChartData with the string of the new locale. One way to solve
this is to watch '$i18n.locale' and trigger fillChartsData when it
changes. My preferred solution is to do away with fillChartsData
completely and make crawledUrlsChartData and totalToolsChartData
computed properties.

Bug: T297860
Change-Id: I2ffd28493485ef7971be8a450432482b4514edb0