api: Convert toolinfo author from string to array of objects
Update database storage and API serialization so that the canonical
representation for the author member of a toolinfo record becomes an
array of Person objects as allowed by the toolinfo.json v1.2.2 schema.

The crawler will continue to support author being provided as a single
string for backwards compatibility with Hay's Directory. When a single
string is provided it is now split on commas and stored as an array of
minimal Person objects (`{name: ...}`).

Database migrations will handle converting existing Tool models to the
new canonical storage format using the same logic as the legacy handling
in the crawler. Historic revisions of Tool models will also be updated
to the new canonical storage format so that they can be deserialized and
rendered by the newer serializer.

When deployed to a dev, testing, or production environment database
migrations must be run followed by a rebuild of the search index:
* `poetry run ./manage.py migrate`
* `poetry run ./manage.py search_index --rebuild`

Bug: T293565
Bug: T301210
Change-Id: I851a205fdfa209b6fda5ff90aa87e6cfeb7f7d70