search: generate Document fields from serializers
Add new helpers which can be used to configure a Document based on
existing serializers created for use with Django REST Framework for API
endpoints. The first use of this functionality is replacing the hand
built ObjectField and schema documentation for the `created_by` member
of a ToolDocument with one generated based on the UserSerializer class
used across many existing API endpoints.

Future patches will use this same functionality to add our new
Annotations information to ToolDocument.

Bug: T299557
Change-Id: If37f35fc08a3cd43dd274b865b1dfecfc6fd6875