annotations: Add permissions
Add permissions to allow all authenticated user to create and edit,
admins to delete, and all to view an annotations model.

In practice, creation and deletion will be handled outside of
permissions verified scopes (creation via post-save trigger for
toolinfo, and deletion via database constraints when a toolinfo is
deleted), but conceptually these are still valid permissions for these
model instances.

Bug: T299557
Change-Id: I52d976c0f9caf40d586989706c7be4bdffc8a189