lists: ToolListDiffSerializer for generating diffs
This change brings back the equivalent of the
ToolListHistoricVersionSerializer that existed before 355e78c as
ToolListDiffSerializer. This serializer outputs each tool in a list as
just the toolname instead of a full object, which is what we need for
comparing diffs of the list without also diffing the tools themselves.

Bug: T300157
Co-Authored-By: Bryan Davis <bd808@wikimedia.org>
Change-Id: I91fa8075f7cbc63e1d3820e974cbefb541705af3