ui: Always cast suffix passed to getMetaTitle as String
The message that is used by getMetaTitle to render the page title uses
a `{{bidi}}` transform. The Banana formatter assumes that any argument
to bidi is a string without guarding against non-string inputs. This
leads to runtime errors when a number is passed. Fix this problem
centrally by always casting a provided suffix value as a String when
building the parameter list for the message.

Change-Id: I184166e2ee37e8412acc9b993c85dc2e581835c1