ui: ensure user is fetched before fetching user's lists
Ensure that information about the current user has been fetched from the
backend before attempting to use the user's username to find lists
created by the user by chaining to the 'user/getUserInfo' vuex action.

The 'user/getUserInfo' action is also updated to make calling easier by
removing the need to pass in the active Vue instance. This was used to
call the $ability plugin. That is now done with the magic `this._vm`
handle to the Vue instance from inside the 'USER' vuex mutation.

Bug: T302467
Change-Id: I45a82ab4d09ce74a6331d2da97913ee672d8d1d0