dev: speed up isValidHttpUrl tests
Improve speed of isValidHttpUrl() as seen during unit tests by moving
the protocol test to a regex combined with the existing unencoded
whitespace guard. This is not expected to have any meaningful impact on
execution speed on the client, but it does greatly improve the speed of
the unit tests which are designed to fail this validation.

Change-Id: I0604dd84d8dede2f61cc8b844220f51985dc1055