dev: enhance mocha reporting
Configure mocha (nodejs unit test framework) to report tests taking more
than 5ms as slow and to enforce an hard timeout of 1 second on any
single test. This is hoped to give us more visibility about which tests
are the most expensive so that we can continue to improve them.

The glob pattern used by stylelint has also been updated to narrow the
top level directories scanned for css and vue files to report on.

Change-Id: Ibfd23a8446840b76e8de685d5e6626b64eccd024