ui: Prevent race condition on calls to api/getRequestSchema
When we have a view with many instances of a component that all try to
fetch the OpenAPI schema at the same time, our current attempts at
caching the first response for subsequent reuse fail.

This change stores the first response as a promise in Vuex for every
other request to use, once resolved.

Bug: T304573
Change-Id: Icf698a780b7c2061e4d71626012ebabc776d97f4