config: Do not force TLS for metrics collection
Add /metrics to the SECURE_REDIRECT_EXEMPT setting to exempt this
endpoint from 301 redirects generated by SecurityMiddleware for non-TLS
requests. The /metrics endpoint is used in Kubernetes deployments to
collect runtime metrics for the live application. The URL is used by
Prometheus data collectors which operate from inside the Kubernetes
cluster and do not traverse the CDN edge which handles TLS termination.

Bug: T306352
Change-Id: I22663d333af7b59dcf3e637fc8b81187c4c463f3