ui(chrome): Change header color to match logo's blue
Change the header color per the "de-emphasize the blue header"
recommendation from past design review. The blue we are now using is the
same blue used in the Toolhub community logo (#069). When combined with
white text (#fff) this has a 6.24:1 contrast ratio. That contrast ratio
exceeds the WCAG AAA standard of 4.5:1 for large text (>=18.66px + bold;
>=24px) and falls just a bit short of the 7:1 standard for normal text.

Bug: T286164
Change-Id: I66ad1ee9ca88e1561e67b1835f73db445556c6fb