login: URLencode query string parameters passed to /social/login/wikimedia/
URLencode the post-authentication continuation data passed from our
login view to the django-social-auth view that handles the OAuth
handshake with MediaWiki. This is important because the continuation may
be a URI with its own query string parameters which inherently means
that it contains characters that are unsafe when left unencoded (`?` and
`&` for example).

Bug: T287179
Change-Id: I962d05fd5bc17e27932e52904a3d4b99a548cbc0