oauth2: Set protocol on custom CSP form-action source
Set an explicit protocol matching the callback URI when extending the
Content-Security-Policy's `form-source` source list.

This is a follow up to 7d77d02. We are still finding issues with
Chromium/Chrome clients rejecting the post-approval redirect when
testing in an environment where the OAuth client is running on localhost
without TLS encryption and the Toolhub instance acting as an OAuth
server is hosted remotely and using TLS on its connection.

Bug: T329563
Change-Id: I7dad84c5bb09e5ef08b2dc98b3442bcb6c2a6c2e