cronjob: Invoke `poetry` as entrypoint without full path
Blubber has changed upstream to install Python libraries in a virtual
environment to make modern Debian base images that mark the system as
`EXTERNALLY-MANAGED` happier. This in turn has changed the location of
the `poetry` entrypoint from the former /usr/local/bin/poetry to
/opt/lib/venv/bin/poetry inside the /opt/lib/venv venv root.

Rather than hard coding the new venv managed location we will now rely
on proper configuration of PATH for the "production" variant's
entrypoint of `poetry ...`.

Bug: T366506
Change-Id: Ib32851defb8caeecc3625282d5a5e3d59a513577